Refactor translateText function
