+++ src/utils/translate.js
@@ -1,0 +1,28 @@
+const axios = require("axios");
+const API_KEY = process.env.GOOGLE_CLOUD_API_KEY;
+
+if (!API_KEY) {
+    console.error("ERROR: GOOGLE_CLOUD_API_KEY is missing in .env file!");
+    process.exit(1);
+}
+
+const translateText = async (text, targetLang) => {
+    try {
+        const response = await axios.post(
+            `https://translation.googleapis.com/language/translate/v2?key=${API_KEY}`,
+            {
+                q: text,
+                source: "en",
+                target: targetLang,
+                format: "text",
+            }
+        );
+
+        return response.data.data.translations[0].translatedText;
+    } catch (err) {
+        console.error(`Translation Error (${targetLang}):`, err.response?.data || err.message);
+        return text;
+    }
+};
+
+module.exports = translateText;
